requirements.txt
pyyaml
psycopg2-binary
python-slugify
sqlalchemy
ckanapi
Flask
python-slugify
click
jenkspy